Update flask_server.py
Appears to be grabbing the artist name and duplicating it for me, so added a test and new variables to accommodate for that, and if the artist name is duped, cut the string in half, use that going forward.
Also changed file extension from mp4 to mp3 so that I didn't have to do any acrobatics under windows to get the file to play in my media player, strictly a personal preference, but made the change anyway.
It seemed like a good idea to put the artist name on the filename, with a separator, for cleanness.